refactor(console): share OSS upsell tracking contracts via schemas
